chat_bot_server/requirements.txt
aioredis==1.3.1
asgiref 
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo
beautifulsoup4==4.6.0
blinker==1.4
brotlipy==0.7.0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2022.5.18.1
cffi 
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer 
constantly==15.1.0
cryptography 
cycler==0.11.0
Cython==0.29.14
daphne==3.0.2
defusedxml 
dj-rest-auth==2.2.4
Django 
django-allauth 
django-firebase-authentication==1.0.9
django-rest-auth
djangorestframework 
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
firebase-admin==5.2.0
Flask==1.1.2
gensim==3.8.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpcio==1.46.1
grpcio-status==1.46.1
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna 
incremental==21.3.0
joblib==0.15.1
JPype1==1.2.0
kiwisolver==1.4.2
kochat==1.0.3
kocrawl==1.0.9
konlpy==0.5.2
lxml==4.8.0
matplotlib==3.2.1
msgpack==1.0.3
numpy==1.18.5
oauthlib 
pandas==1.0.4
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
PyJWT==2.3.0
PyMySQL 
pyngrok 
pyOpenSSL
pyparsing==3.0.8
PySocks 
python-dateutil==2.8.2
python3-openid 
pytorch-crf==0.7.2
pytz 
PyYAML==6.0
requests 
requests-oauthlib
rsa==4.8
scikit-learn==0.23.1
scipy==1.8.0
service-identity==21.1.0
six 
smart-open==6.0.0
sqlparse 
threadpoolctl==3.1.0
torch==1.11.0
tqdm==4.64.0
tweepy==3.10.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions
tzdata 
uritemplate==4.1.1
urllib3 
win-inet-pton
wincertstore==0.2
zope.interface==5.4.0
